requirements.txt
asgiref==3.3.4
Brotli==1.0.9
cffi==1.14.6
cssselect2==0.4.1
Django==3.2.4
django-filter==2.4.0
fonttools==4.25.2
html5lib==1.1
nepali-date==2.0.5
nepali-datetime==1.0.7
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pydyf==0.0.3
pyphen==0.11.0
pytz==2021.1
reportlab==3.5.68
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
weasyprint==53.0
webencodings==0.5.1
zopfli==0.1.8
